Deprecate options.dart.
