+++ test/warnings_test.dart
@@ -6,7 +6,6 @@
 library dartdoc.warnings_test;
 
 import 'package:analyzer/file_system/physical_file_system.dart';
-import 'package:dartdoc/options.dart';
 import 'package:dartdoc/src/dartdoc_options.dart';
 import 'package:dartdoc/src/generator/generator.dart';
 import 'package:dartdoc/src/io_utils.dart';
